requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow
watchdog

